% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{calculate_activity}
\alias{calculate_activity}
\title{Calculate Activity Index Based on Zone Transitions}
\usage{
calculate_activity(df)
}
\arguments{
\item{df}{A data frame with columns \code{rfid}, \code{datetime}, and \code{zone}}
}
\value{
The input data frame with an added \code{activity_index} column (in inches).
}
\description{
Computes theoretical distance traveled between sequential zone visits
using UID Mouse Matrix coordinates. Adds a new column \code{activity_index}
to the data frame.
}
