% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bouts.R
\name{compute_bout_durations}
\alias{compute_bout_durations}
\title{Compute bout durations}
\usage{
compute_bout_durations(
  df,
  sampling_interval,
  duration_mode = c("strict", "inclusive")
)
}
\arguments{
\item{df}{A grouped df with \code{start} and \code{end} timestamps per bout.}

\item{sampling_interval}{Numeric value representing sampling interval in minutes.}

\item{duration_mode}{Either "strict" (end - start) or "inclusive" (end - start + sampling_interval).}
}
\value{
A data frame with a \code{duration_minutes} column.
}
\description{
Compute bout durations
}
\keyword{internal}
