% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{downsample_activity}
\alias{downsample_activity}
\title{Downsample activity index data to fixed intervals}
\usage{
downsample_activity(df, n = 1, precision = "minute")
}
\arguments{
\item{df}{A cleaned dataframe with \code{datetime}, \code{rfid}, \code{session_name}, \code{matrix_name}, and \code{activity_index}.}

\item{n}{Number of time units per bin (default = 1).}

\item{precision}{Time unit for binning (e.g., "minute").}
}
\value{
Downsampled data frame with activity index summarized by total distance per bin. Missing data for time bins will be filled with \code{NA}.
}
\description{
Downsample activity index data to fixed intervals
}
\seealso{
\code{\link[=calculate_activity]{calculate_activity()}}
}
