% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_temperature.R
\name{downsample_temperature}
\alias{downsample_temperature}
\title{Downsample temperature data to fixed intervals}
\usage{
downsample_temperature(df, n = 1, precision = "minute")
}
\arguments{
\item{df}{A cleaned dataframe with \code{datetime}, \code{rfid}, \code{session_name}, \code{matrix_name}, and \code{temperature}.}

\item{n}{Number of time units per bin (default = 1).}

\item{precision}{Time unit for binning (e.g., "minute").}
}
\value{
Downsampled data frame with temperature summarized by median. Missing data for time bins will be filled with \code{NA}.
}
\description{
Downsample temperature data to fixed intervals
}
