% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bouts.R
\name{estimate_sampling_interval}
\alias{estimate_sampling_interval}
\title{Estimate Sampling Interval from Grouped Data}
\usage{
estimate_sampling_interval(df)
}
\arguments{
\item{df}{A grouped data frame containing a \code{common_dt} column.}
}
\value{
Numeric sampling interval in minutes.
}
\description{
Determines the time difference in minutes between consecutive \code{common_dt}
observations for each group in \code{df}. All groups must have a single unique
interval and share the same value.
}
