% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_temperature.R
\name{find_raw_export_files}
\alias{find_raw_export_files}
\title{Finds raw exported files ending with CSV}
\usage{
find_raw_export_files(raw_export_dir)
}
\arguments{
\item{raw_export_dir}{Path to the directory where raw data is located.}
}
\description{
Finds raw exported files ending with CSV
}
\keyword{internal}
