% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_temperature.R
\name{flag_temperature_outliers}
\alias{flag_temperature_outliers}
\title{Flag temperature outliers based on rolling difference threshold}
\usage{
flag_temperature_outliers(df, threshold = 1)
}
\arguments{
\item{df}{A dataframe with \code{datetime}, \code{rfid}, and \code{temperature}.}

\item{threshold}{Maximum allowed temperature jump (default = 1).}
}
\value{
Data frame with added \code{temp_diff} and \code{outlier_global} columns.
}
\description{
Flag temperature outliers based on rolling difference threshold
}
