% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_temperature.R
\name{generate_cleaning_plots}
\alias{generate_cleaning_plots}
\title{Generate diagnostic plots for outliers and downsampled data}
\usage{
generate_cleaning_plots(
  df_outlier_flagged,
  df_downsampled,
  filepath,
  output_dir
)
}
\arguments{
\item{df_outlier_flagged}{Data frame with outlier flags.}

\item{df_downsampled}{Downsampled temperature data.}

\item{filepath}{Path to the original file (used for naming output plots).}

\item{output_dir}{Where to save the plots.}
}
\description{
Generate diagnostic plots for outliers and downsampled data
}
