% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_processing.R
\name{pluck_bind_temp_activity}
\alias{pluck_bind_temp_activity}
\title{Helper to pluck the temperature and activity datasets separately
we can consider including this instead of the bind_rows(.id = "variable")
the con of this is handling multiple files}
\usage{
pluck_bind_temp_activity(clean_list, nm)
}
\description{
Helper to pluck the temperature and activity datasets separately
we can consider including this instead of the bind_rows(.id = "variable")
the con of this is handling multiple files
}
\keyword{internal}
